doing homework from lecture9

+++ src/Lecture8/exercise2.java
@@ -2,14 +2,14 @@ package Lecture8;
 
 public class exercise2 {
     public static void main(String[] args) {
-        int a = 30;
-        int b = 10;
-        int c = 40;
+        double a = 30;
+        double b = 10;
+        double c = 40;
         System.out.println("The average sum is: " + averageSum(a,b,c));
     }
 
-    public static double averageSum(int a, int b, int c){
-        double averageSum = (a + b + c) / 3.0;
+    public static double averageSum(double a, double b, double c){
+        double averageSum = (a + b + c) / 3;
         return averageSum;
     }
 }
